add param add_root_ca
This parameter is used to assign an enterprise certification authority (CA).
Specify the entire certificate with the designation of the beginning and end of the certificate,
and can be without the designation of the beginning and end "-----BEGIN CERTIFICATE-----", "-----END CERTIFICATE-----".
In a regular expression, these parts of the strings are specified as is and any change in these parts will result in a certificate installation error!
In the file mozilla.cfg is assigned to the variable as a whole string without spaces and hyphenation(otherwise it does not work).